refactor: unify converting arcade-record-posts to search-result-props
