+++ src/entities/search/types.ts
@@ -1,0 +1,8 @@
+export interface SearchResultProps {
+  title: string;
+  subheading?: string;
+  description: string;
+  href: string;
+  thumbnailUrl: string;
+  emphasize: string;
+}
